+++ lib/neatogen/adjust.c
@@ -569,8 +569,8 @@ SparseMatrix makeMatrix(Agraph_t *g) {
 	}
     }
 
-    SparseMatrix A = SparseMatrix_from_coordinate_arrays(nedges, nnodes, nnodes,
-                                                         I, J, val,
+    SparseMatrix A = SparseMatrix_from_coordinate_arrays((size_t)nedges, nnodes,
+                                                         nnodes, I, J, val,
                                                          MATRIX_TYPE_REAL,
                                                          sizeof(double));
